separate the dom part of the map js with the plugin part of the map js
