+++ app/javascript/dom/index.js
@@ -1,1 +1,2 @@
 import './toggler'
+import './map-filter'
